normed , segfault to fix

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: lweglarz <lweglarz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/10 12:34:54 by lweglarz          #+#    #+#             */
-/*   Updated: 2020/08/13 14:23:56 by lweglarz         ###   ########.fr       */
+/*   Updated: 2020/08/19 16:25:51 by lweglarz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,8 +22,8 @@ int		get_next_line(int fd, char **line);
 
 char	*ft_substr(char const *s, unsigned int start, size_t len);
 char	*ft_strdup(const char *s);
-int		ft_strlenc(const char *str, char c);
-char		*ft_strjoin(char const *s1, char const *s2);
+size_t	ft_strlenc(const char *str, char c);
+char	*ft_strjoin(char const *s1, char const *s2);
 char	*ft_strchr(const char *s, int c);
 
 #endif
